http service and group component

+++ src/app/auth/services/auth.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
 import { tap } from 'rxjs/operators';
 import { BehaviorSubject } from 'rxjs';
 import { User } from '../user.model';
 import { Router } from '@angular/router';
 import { StorageService } from '../../services/storage.service';
+import { HttpService } from 'src/app/services/http.service';
 
 /*
 this service will handle all login , logout , register operations for the whole app
@@ -12,29 +12,23 @@ and contains a user subject to subscribe to any time for checking user status
 logged in or not
 */
 
-const authEndPoints = {
-  csrf: 'http://localhost:8000/sanctum/csrf-cookie',
-  login: 'http://localhost:8000/api/v1/login',
-  register: '',
-};
-
 @Injectable({
   providedIn: 'root',
 })
 export class AuthService {
   user = new BehaviorSubject<User>(null);
   constructor(
-    private http: HttpClient,
+    private httpService: HttpService,
     private router: Router,
     private storageService: StorageService
   ) {}
 
   login(email: string, password: string) {
-    return this.http
-      .post(authEndPoints.login, {
+    return this.httpService
+      .requestLogin({
         email: email,
         password: password,
-        device_name: 'test',
+        device_name: navigator.platform,
       })
       .pipe(
         tap((res: any) => {
@@ -47,6 +41,7 @@ export class AuthService {
             user.address,
             user.mobile,
             user.avatar,
+            user.verified,
             token
           );
           this.user.next(currentUser);
@@ -65,8 +60,18 @@ export class AuthService {
   }
 
   logout() {
-    this.user.next(null);
-    this.storageService.removeItem('user');
-    this.router.navigate(['/login']);
+    this.httpService.requestLogout(navigator.platform).subscribe(
+      (res) => {
+        if (res) {
+          console.log(res);
+          this.user.next(null);
+          this.storageService.removeItem('user');
+          this.router.navigate(['/login']);
+        }
+      },
+      (error) => {
+        console.log(error);
+      }
+    );
   }
 }
